all tests pass on the back end

+++ server/src/services/CommentsService.js
@@ -17,14 +17,14 @@ class CommentsService {
   }
 
   async deleteComment(commentId, userId) {
-    const commentToDelete = await dbContext.Comments.findById(commentId, userId).populate('creator')
+    const commentToDelete = await dbContext.Comments.findById(commentId).populate('creator')
     if (!commentToDelete) {
       throw new Error('Wait, there is no comment to delete')
     }
 
-    // if (commentToDelete.creatorId != userId) {
-    //   throw new Forbidden('oops, you are not authorized to delete this')
-    // }
+    if (commentToDelete.creatorId != userId) {
+      throw new Forbidden('oops, you are not authorized to delete this')
+    }
 
     await commentToDelete.deleteOne()
     return 'comment was removed'
